asset submit

+++ src/main/java/com/fiveonechain/digitasset/controller/UserController.java
@@ -176,7 +176,7 @@ public class UserController {
             Result result = ResultUtil.buildErrorResult(ErrorInfo.TELEPHONE_VERIFY_FAIL);
             return result;
         }
-        User user = iUserService.getUserByUserId((long) userContext.getUserId());
+        User user = iUserService.getUserByUserId(userContext.getUserId());
         user.setTelephone(telephone);
         if (!iUserService.updateMobile(user)) {
             Result result = ResultUtil.buildErrorResult(ErrorInfo.SERVER_ERROR);
@@ -307,7 +307,7 @@ public class UserController {
             Result result = ResultUtil.buildErrorResult(ErrorInfo.IMAGE_EXISTS);
             return result;
         }
-        User user = iUserService.getUserByUserId(Long.valueOf(userContext.getUserId()));
+        User user = iUserService.getUserByUserId(userContext.getUserId());
         if (user == null) {
             Result result = ResultUtil.buildErrorResult(ErrorInfo.USER_NOT_FOUND);
             return result;
@@ -359,7 +359,7 @@ public class UserController {
             Result result = ResultUtil.buildErrorResult(ErrorInfo.CORP_EXISTS);
             return result;
         }
-        User user = iUserService.getUserByUserId(Long.valueOf(userContext.getUserId()));
+        User user = iUserService.getUserByUserId(userContext.getUserId());
         if (user == null) {
             Result result = ResultUtil.buildErrorResult(ErrorInfo.USER_NOT_FOUND);
             return result;
@@ -386,7 +386,7 @@ public class UserController {
     public Result getCorpDetail(@AuthenticationPrincipal UserContext userContext
     ) {
         //验证identity后设置status
-        User user = iUserService.getUserByUserId(Long.valueOf(userContext.getUserId()));
+        User user = iUserService.getUserByUserId(userContext.getUserId());
         if (user == null) {
             Result result = ResultUtil.buildErrorResult(ErrorInfo.USER_NOT_FOUND);
             return result;
