+++ src/main/java/com/fiveonechain/digitasset/domain/User.java
@@ -7,8 +7,8 @@ import com.fasterxml.jackson.annotation.JsonIgnore;
  */
 public class User {
 
-    private Long id;
-    private Long user_id;
+    private Integer id;
+    private Integer user_id;
     private String user_name;
     @JsonIgnore
     private String password;
@@ -29,19 +29,19 @@ public class User {
         this.token = token;
     }
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
 
-    public Long getUser_id() {
+    public Integer getUser_id() {
         return user_id;
     }
 
-    public void setUser_id(Long user_id) {
+    public void setUser_id(Integer user_id) {
         this.user_id = user_id;
     }
